fix: error when running tests
The spec we pass to lazy.minit.setup has `opts={}`.  This caused an
error because LazyVim would load the plugin and try to call `setup` but
our module had no so function.

To avoid having setup called one must pass `opts = nil`, but this commit
instead adds a no-op setup function since that's closer to what we'll
want out of a shared functionality plugin.